EPMEDU-1079 fix detekt notes
