+++ app/src/main/java/com/epmedu/animeal/component/DataStorePreferences.kt
@@ -6,14 +6,11 @@ import androidx.datastore.preferences.core.booleanPreferencesKey
 import androidx.datastore.preferences.core.stringPreferencesKey
 import com.epmedu.animeal.common.constants.DefaultConstants
 
-private val ANIMAL_TYPE  = stringPreferencesKey("AnimalType")
+private val ANIMAL_TYPE = stringPreferencesKey("AnimalType")
 val Preferences.animalType: String
     get() = this[ANIMAL_TYPE] ?: DefaultConstants.EMPTY_STRING
 fun MutablePreferences.updateAnimalType(value: String) {
     this[ANIMAL_TYPE] = value
-}
-fun MutablePreferences.clearAnimalType() {
-    this.remove(ANIMAL_TYPE)
 }
 
 private val INITIAL_GEOLOCATION_PERMISSION = booleanPreferencesKey("InitialGeolocationPermissionOnHomeScreen")
@@ -22,16 +19,10 @@ val Preferences.initialGeolocationPermission: Boolean
 fun MutablePreferences.updateInitialGeolocationPermission(value: Boolean) {
     this[INITIAL_GEOLOCATION_PERMISSION] = value
 }
-fun MutablePreferences.clearInitialGeolocationPermission() {
-    this.remove(INITIAL_GEOLOCATION_PERMISSION)
-}
 
-private val INITIAL_CAMERA_PERMISSION  = booleanPreferencesKey("InitialCameraPermissionOnHomeScreen")
+private val INITIAL_CAMERA_PERMISSION = booleanPreferencesKey("InitialCameraPermissionOnHomeScreen")
 val Preferences.initialCameraPermission: Boolean
     get() = this[INITIAL_CAMERA_PERMISSION] ?: false
 fun MutablePreferences.updateInitialCameraPermission(value: Boolean) {
     this[INITIAL_CAMERA_PERMISSION] = value
-}
-fun MutablePreferences.clearInitialCameraPermission() {
-    this.remove(INITIAL_CAMERA_PERMISSION)
 }
